+++ examples/bench_ieee39.rs
@@ -1,3 +1,4 @@
+#![allow(deprecated)]
 use rustpower::{io::pandapower::Network, prelude::*};
 
 #[macro_export]
