+++ examples/ieee118.rs
@@ -1,3 +1,4 @@
+#![allow(deprecated)]
 use std::env;
 
 use nalgebra::ComplexField;
