+++ src/basic/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod dsbus_dv;
 pub(crate) mod newtonpf;
 
-pub mod new_ecs;
+pub mod ecs;
 pub mod solver;
 pub(crate) mod sparse;
 pub mod system;
