+++ src/io/pandapower/network_converter.rs
@@ -1,4 +1,6 @@
+#![allow(deprecated)]
 use std::f64::consts::PI;
+
 
 use crate::basic::system::*;
 use crate::prelude::pandapower::*;
